move all authentication and sso related cruft into a controller concern (#2634)
* move all authentication and sso related cruft into a controller concern

* forgot to add the concern

* minor tweaks

* prepend certain before actions for block unknown hosts and set app info headers to occur before authN

* oops

* rubocop
