+++ app/controllers/application_controller.rb
@@ -8,8 +8,7 @@ require 'sentry_logging'
 require 'aes_256_cbc_encryptor'
 
 class ApplicationController < ActionController::API
-  include ActionController::HttpAuthentication::Token::ControllerMethods
-  include ActionController::Cookies
+  include AuthenticationAndSSOConcerns
   include SentryLogging
   include Pundit
 
@@ -20,17 +19,10 @@ class ApplicationController < ActionController::API
     Breakers::OutageException
   ].freeze
 
-  before_action :block_unknown_hosts
-  before_action :authenticate
-
-  # Ensures that we maintain sessions for currently signed-in users
-  # (that have been authenticated with the HTTP header)
-  # after we start using cookies instead of the header.
-  before_action :set_api_cookie!, unless: -> { Settings.session_cookie.enabled }
-
-  before_action :set_app_info_headers
-  before_action :set_tags_and_extra_context
+  prepend_before_action :block_unknown_hosts, :set_app_info_headers
+  # See Also AuthenticationAndSSOConcerns for more before filters
   skip_before_action :authenticate, only: %i[cors_preflight routing_error]
+  before_action :set_tags_and_extra_context
 
   def tag_rainbows
     Sentry::TagRainbows.tag
@@ -40,12 +32,12 @@ class ApplicationController < ActionController::API
     head(:ok)
   end
 
-  def clear_saved_form(form_id)
-    InProgressForm.form_for_user(form_id, @current_user)&.destroy if @current_user
-  end
-
   def routing_error
     raise Common::Exceptions::RoutingError, params[:path]
+  end
+
+  def clear_saved_form(form_id)
+    InProgressForm.form_for_user(form_id, @current_user)&.destroy if @current_user
   end
 
   # I'm commenting this out for now, we can put it back in if we encounter it
@@ -142,93 +134,6 @@ class ApplicationController < ActionController::API
     headers['X-Git-SHA'] = AppInfo::GIT_REVISION
   end
 
-  def authenticate
-    authenticate_token || render_unauthorized
-  end
-
-  def authenticate_token
-    return validate_session(session[:token]) if Settings.session_cookie.enabled
-    authenticate_with_http_token do |token, _options|
-      validate_session(token)
-    end
-  end
-
-  def validate_session(token)
-    @session_object = Session.find(token)
-
-    if @session_object.nil?
-      Rails.logger.info('SSO: INVALID SESSION', sso_logging_info)
-      reset_session
-      return false
-    end
-
-    @current_user = User.find(@session_object.uuid)
-
-    if should_signout_sso?
-      Rails.logger.info('SSO: MHV INITIATED SIGNOUT', sso_logging_info)
-      reset_session
-    else
-      Rails.logger.info('SSO: EXTENDING SESSION', sso_logging_info)
-      extend_session!
-    end
-
-    @current_user.present?
-  end
-
-  def extend_session!
-    Rails.logger.info('SSO: ApplicationController#extend_session!', sso_logging_info)
-
-    @session_object.expire(Session.redis_namespace_ttl)
-    @current_user&.identity&.expire(UserIdentity.redis_namespace_ttl)
-    @current_user&.expire(User.redis_namespace_ttl)
-    set_sso_cookie!
-  end
-
-  def reset_session
-    Rails.logger.info('SSO: ApplicationController#reset_session', sso_logging_info)
-
-    cookies.delete(Settings.sso.cookie_name, domain: Settings.sso.cookie_domain)
-    @session_object&.destroy
-    @current_user&.destroy
-    @session_object = nil
-    @current_user = nil
-    super
-  end
-
-  # Sets a cookie "api_session" with all of the key/value pairs from session object.
-  def set_api_cookie!
-    return unless @session_object
-    @session_object.to_hash.each { |k, v| session[k] = v }
-  end
-
-  # https://github.com/department-of-veterans-affairs/vets.gov-team/blob/master/Products/SSO/CookieSpecs-20180906.docx
-  def set_sso_cookie!
-    Rails.logger.info('SSO: ApplicationController#set_sso_cookie!', sso_logging_info)
-
-    return unless Settings.sso.cookie_enabled && @session_object.present?
-    encryptor = SSOEncryptor
-    contents = ActiveSupport::JSON.encode(@session_object.cookie_data(@current_user))
-    encrypted_value = encryptor.encrypt(contents)
-    cookies[Settings.sso.cookie_name] = {
-      value: encrypted_value,
-      expires: nil, # NOTE: we track expiration as an attribute in "value." nil here means kill cookie on browser close.
-      secure: Settings.sso.cookie_secure,
-      httponly: true,
-      domain: Settings.sso.cookie_domain
-    }
-  end
-
-  def should_signout_sso?
-    Rails.logger.info('SSO: ApplicationController#should_signout_sso?', sso_logging_info)
-    return false unless Settings.sso.cookie_enabled
-    return false unless Settings.sso.cookie_signout_enabled
-    cookies[Settings.sso.cookie_name].blank? && request.host.match(Settings.sso.cookie_domain)
-  end
-
-  def render_unauthorized
-    raise Common::Exceptions::Unauthorized
-  end
-
   def saml_settings(options = {})
     callback_url = URI.parse(Settings.saml.callback_url)
     callback_url.host = request.host
@@ -245,15 +150,5 @@ class ApplicationController < ActionController::API
 
   def render_job_id(jid)
     render json: { job_id: jid }, status: 202
-  end
-
-  def sso_logging_info
-    {
-      sso_cookies_enabled: Settings.sso.cookie_enabled,
-      sso_cookies_signout_enabled: Settings.sso.cookie_signout_enabled,
-      sso_cookie_name: Settings.sso.cookie_name,
-      sso_cookie_contents: @session_object&.cookie_data(@current_user),
-      request_host: request.host
-    }
   end
 end
