+++ app/models/session.rb
@@ -34,17 +34,6 @@ class Session < Common::RedisStore
     super(ttl)
   end
 
-  # https://github.com/department-of-veterans-affairs/vets.gov-team/blob/master/Products/SSO/CookieSpecs-20180906.docx
-  def cookie_data(user)
-    return nil if user.blank?
-    raise 'Invalid User UUID' unless user.uuid == uuid
-    {
-      'patientIcn' => user.mhv_icn || user.icn,
-      'mhvCorrelationId' => user.mhv_correlation_id,
-      'expirationTime' => ttl_in_time.iso8601(0)
-    }
-  end
-
   def ttl_in_time
     Time.current.utc + ttl
   end
